реализована 2х-уровневая архитектура для работы с группами
