+++ fixture/application.py
@@ -1,44 +1,19 @@
 from selenium.webdriver.chrome.webdriver import WebDriver
 from fixture.session import SessionHelper
+from fixture.group import GroupHelper
 
 
 class Application:
 
     def __init__(self):
-        self.wd = WebDriver()
+        self.wd = WebDriver()    # ссылка на веб-драйвер
         self.wd.implicitly_wait(30)
         self.session = SessionHelper(self)
+        self.group = GroupHelper(self)
 
     def open_home_page(self):
         wd = self.wd
         wd.get("http://localhost/addressbook/")
-
-    def open_group_page(self):
-        wd = self.wd
-        wd.find_element_by_link_text("groups").click()
-
-    def create_new_group(self, group_param):
-        wd = self.wd
-        self.open_group_page()
-        # init new creation
-        wd.find_element_by_name("new").click()
-        # fill group form
-        wd.find_element_by_name("group_name").click()
-        wd.find_element_by_name("group_name").clear()
-        wd.find_element_by_name("group_name").send_keys(group_param.group_name)
-        wd.find_element_by_name("group_header").click()
-        wd.find_element_by_name("group_header").clear()
-        wd.find_element_by_name("group_header").send_keys(group_param.header)
-        wd.find_element_by_name("group_footer").click()
-        wd.find_element_by_name("group_footer").clear()
-        wd.find_element_by_name("group_footer").send_keys(group_param.footer)
-        # submit new creation
-        wd.find_element_by_name("submit").click()
-        self.return_to_group_page()
-
-    def return_to_group_page(self):
-        wd = self.wd
-        wd.find_element_by_link_text("group page").click()
 
     def destroy(self):
         self.wd.quit()
