+++ fixture/group.py
@@ -1,0 +1,31 @@
+class GroupHelper:
+    def __init__(self, app):
+        self.app = app    # ссылка на объект класса Application
+
+    def open_group_page(self):
+        wd = self.app.wd
+        wd.find_element_by_link_text("groups").click()
+
+    def create(self, group_param):
+        wd = self.app.wd
+        self.open_group_page()
+        # init new creation
+        wd.find_element_by_name("new").click()
+        # fill group form
+        wd.find_element_by_name("group_name").click()
+        wd.find_element_by_name("group_name").clear()
+        wd.find_element_by_name("group_name").send_keys(group_param.group_name)
+        wd.find_element_by_name("group_header").click()
+        wd.find_element_by_name("group_header").clear()
+        wd.find_element_by_name("group_header").send_keys(group_param.header)
+        wd.find_element_by_name("group_footer").click()
+        wd.find_element_by_name("group_footer").clear()
+        wd.find_element_by_name("group_footer").send_keys(group_param.footer)
+        # submit new creation
+        wd.find_element_by_name("submit").click()
+        self.return_to_group_page()
+
+    def return_to_group_page(self):
+        wd = self.app.wd
+        wd.find_element_by_link_text("group page").click()
+
