+++ model/group_param.py
@@ -3,7 +3,3 @@ class GroupParam:
         self.group_name = group_name
         self.header = header
         self.footer = footer
-
-
-# a = GroupParam(1, 2, 3)
-# print(a.group_name, a.header, a.footer)
